+++ common/src/main/java/com/railwayteam/railways/compat/Mods.java
@@ -29,7 +29,6 @@ import java.util.function.Supplier;
  * For compatibility with and without another mod present, we have to define load conditions of the specific code
  */
 public enum Mods {
-	JOURNEYMAP("journeymap-fabric"),
 	BIOMESOPLENTY("biomesoplenty"),
 	BLUE_SKIES("blue_skies"),
 	BYG("byg"),