+++ common/src/main/java/com/railwayteam/railways/registry/CRCommandsClient.java
@@ -26,7 +26,6 @@ import com.railwayteam.railways.registry.commands.ClearCapCacheCommand;
 import com.railwayteam.railways.registry.commands.ClearCasingCacheCommand;
 import com.railwayteam.railways.registry.commands.MixinAuditCommand;
 import com.railwayteam.railways.registry.commands.ReloadDevCapesCommand;
-import com.railwayteam.railways.registry.commands.ReloadJourneymapCommand;
 import com.railwayteam.railways.util.Utils;
 import net.minecraft.commands.SharedSuggestionProvider;
 
@@ -42,7 +41,6 @@ public class CRCommandsClient {
             .requires(cs -> cs.hasPermission(0))
             .then(ClearCasingCacheCommand.register())
             .then(ClearCapCacheCommand.register())
-            .then(ReloadJourneymapCommand.register())
             .then(ReloadDevCapesCommand.register());
 
         if (Utils.isDevEnv() || FORCE_MIXIN_AUDIT_COMMAND) {